Setting --verbose flag with no argument causes segfault
On OSX 10.11 when the --verbose flag is set without specifying an argument,
the ConfigManager tries to extract an argument from where one would normally
be found (the extern "optarg" variable). When no argument is actually present,
"optarg" points to garbage data which is read in as a string by atoi() and causes
a segfault. Fixed by changing --verbose to optional, which causes optarg to be
set to zero if no argument is given. This can then be detected and a default value
of 1 supplied.