OSX 10.11 OpenAL Loading Failure
Normally, OpenAL can be loaded on OSX by just loading the string "OpenAL".
For whatever reason, on my copy of 10.11 with everything set up correctly as
far as I can tell, this doesn't work. However, specifying the path to the library
explicitly does work fine. This commit adds an optional define that can be
specified to cmake that will override the OpenAL library loading string on OSX
and allow you to specify the path to OpenAL explicitly. By default, the old
method is used if not specified.